separate claude, batch_bench, prompt_crafter tests
