+++ assets/aml-benchmark/components/batch-inference-preparer/asset.yaml
@@ -5,4 +5,4 @@ test:
   pytest:
     enabled: true
     conda_environment: ../../dev_conda_env.yaml
-    tests_dir: ../../tests
+    tests_dir: ../../tests/test_batch_benchmark_inference
