+++ assets/aml-benchmark/components/batch_benchmark_inference_claude/asset.yaml
@@ -5,4 +5,4 @@ test:
   pytest:
     enabled: true
     conda_environment: ../../dev_conda_env.yaml
-    tests_dir: ../../tests
+    tests_dir: ../../tests/test_claude
