+++ assets/aml-benchmark/tests/test_batch_benchmark_inference/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) Microsoft Corporation.
+# Licensed under the MIT License.
+
+"""Init file."""
