+++ assets/aml-benchmark/tests/test_batch_benchmark_inference/test_batch_benchmark_inference.py
@@ -14,7 +14,7 @@ import pytest
 from azure.ai.ml.entities import Job
 from azure.ai.ml import Input
 
-from .test_utils import (
+from ..test_utils import (
     get_mlclient,
     Constants,
     download_outputs,
