+++ assets/aml-benchmark/tests/test_batch_benchmark_inference/test_batch_benchmark_score.py
@@ -12,7 +12,7 @@ import pytest
 from azure.ai.ml.entities import Job
 from azure.ai.ml import Input
 
-from .test_utils import (
+from ..test_utils import (
     load_yaml_pipeline,
     get_mlclient,
     Constants,
