+++ assets/aml-benchmark/tests/test_prompt_crafter/test_checksum.py
@@ -6,7 +6,7 @@
 import sys
 import pytest
 
-from .test_utils import get_src_dir
+from ..test_utils import get_src_dir
 
 sys.path.append(get_src_dir())
 try:
