+++ assets/aml-benchmark/tests/test_prompt_crafter/test_prompt_crafter_checksum.py
@@ -8,7 +8,7 @@ import sys
 import pytest
 import mlflow
 
-from .test_utils import get_src_dir
+from ..test_utils import get_src_dir
 
 sys.path.append(get_src_dir())
 try:
