+++ assets/aml-benchmark/tests/test_prompt_crafter/test_prompts.py
@@ -7,7 +7,7 @@ import sys
 import pytest
 from typing import List, Dict
 
-from .test_utils import get_src_dir
+from ..test_utils import get_src_dir
 
 sys.path.append(get_src_dir())
 try:
